@@ -1,25 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using MUDService.Models;
-using Microsoft.EntityFrameworkCore;
+using MUDService.DataAccess;
 using MUDService.Helpers;
 using MUDService.Hubs;
-using MUDService.DataAccess;
-using Microsoft.AspNetCore.SignalR;
-using Serilog;
+using MUDService.Models;
 
 namespace MUDService
 {
     public class Startup
     {
-        public Startup(IHostingEnvironment env, IConfiguration configuration)
+        public Startup(IWebHostEnvironment env, IConfiguration configuration)
         {
             var builder = new ConfigurationBuilder()
                 .SetBasePath(env.ContentRootPath)
@@ -28,10 +22,6 @@ namespace MUDService
                 .AddEnvironmentVariables();
             Configuration = builder.Build();
             Configuration = configuration;
-            Log.Logger = new LoggerConfiguration()
-                .Enrich.FromLogContext()
-                .WriteTo.Console()
-                .CreateLogger();
         }
 
         public IConfiguration Configuration { get; }
@@ -39,9 +29,6 @@ namespace MUDService
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddLogging(loggingBuilder =>
-          loggingBuilder.AddSerilog(dispose: true));
-
             services.AddDbContextPool<MUDContext>(options =>
                 options.UseSqlServer(Configuration.GetConnectionString("MUDContext")));
 
@@ -55,35 +42,13 @@ namespace MUDService
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
         {
-            loggerFactory.AddConsole(Configuration.GetSection("Logging"));
-            loggerFactory.AddSerilog();
-            
-            if (env.IsDevelopment())
+            app.UseEndpoints(endpoints =>
             {
-                loggerFactory.AddDebug();
-
-                loggerFactory.AddFile("log.txt");
-            }
-            else
-            {
-                loggerFactory.AddDebug();
-
-                loggerFactory.AddFile("log.txt");
-            }
-
-            //if (env.IsDevelopment())
-            //{
-
-            //}
-
-            app.UseSignalR(routes =>
-            {
-                routes.MapHub<MUDHub>("/MUDHub");
+                endpoints.MapHub<MUDHub>("/MUDHub");
+                endpoints.MapControllers();
             });
-
-            app.UseMvcWithDefaultRoute();
         }
     }
 }
