@@ -1,33 +1,73 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
+using MUDService.DataAccess;
 using MUDService.JsonModels;
+using MUDService.Models;
 using MUDService.ViewModels;
-using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
 
 namespace MUDService.Controllers
 {
     [Route("[controller]")]
     public class WorldImportController : Controller
     {
+        private readonly MUDDataAccess _mudDataAccess;
+        public WorldImportController(MUDDataAccess mudDataAccess)
+        {
+            _mudDataAccess = mudDataAccess;
+        }
         [HttpGet]
         public IActionResult Index()
         {
             var model = new WorldImportViewModel();
-            model.WorldDefinition = "{\"Name\" : \"Test\"}";
-            var jWorld = JsonConvert.DeserializeObject<JWorld>(model.WorldDefinition);
+            model.WorldDefinition = new JWorld()
+            {
+                Name = "Test"
+            };
             return View(model);
         }
 
         [HttpPost]
-        public IActionResult Import(WorldImportViewModel viewModel)
+        public IActionResult Import(WorldImportViewModel request)
         {
-            var jsonString = viewModel.WorldDefinition;
+            var world = _mudDataAccess.CreateNewWorld(request.WorldDefinition.Name);
+            foreach (var jCell in request.WorldDefinition.Cells)
+            {
+                var entities = new List<Entity>();
 
-            var jWorld = JsonConvert.DeserializeObject<JWorld>(jsonString);
-            return null;
+                foreach (var jEntity in jCell.Entities)
+                {
+                    var entityId = Guid.NewGuid();
+                    switch (jEntity.Type)
+                    {
+                        case Entity.EntityType.Item:
+                            var item = new Item()
+                            {
+                                EntityId = entityId,
+                                Type = Entity.EntityType.Item,
+                                Name = jEntity.Name,
+                                Description = jEntity.Description,
+                                Effects = new List<Effect>() //not going to worry about this yet                                
+                            };
+                            entities.Add(item);
+                            break;
+                        default:
+                            var entity = new Entity()
+                            {
+                                EntityId = entityId,
+                                Name = jEntity.Name,
+                                Description = jEntity.Description,
+                                Type = Entity.EntityType.None
+                            };
+                            entities.Add(entity);
+                            break;
+                    }
+                }
+
+                var cell = _mudDataAccess.CreateNewCell(world, jCell.X, jCell.Y, jCell.Z, jCell.Description, entities);
+            }
+
+            return View();
         }
     }
 }
