@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using static MUDService.Models.Entity;
 
 namespace MUDService.JsonModels
 {
@@ -9,6 +6,6 @@ namespace MUDService.JsonModels
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Type { get; set; }
+        public EntityType Type { get; set; }
     }
 }
