@@ -1,13 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace MUDService.JsonModels
 {
     public class JWorld
     {
         public string Name { get; set; }
+        public string Description { get; set; }
         public List<JCell> Cells { get; set; }
     }
 }
