@@ -1,13 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace MUDService.Models
 {
     public class Player : Character
     {
-
         public List<ChatUser> ChatUsers { get; set; }
     }
 }
