@@ -1,15 +1,11 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using MUDService.JsonModels;
 
 namespace MUDService.ViewModels
 {
-    
+
     public class WorldImportViewModel
     {
-        [FromForm]
-        public string WorldDefinition { get; set; }
+        
+        public JWorld WorldDefinition { get; set; }
     }
 }
