@@ -7,6 +7,6 @@
 <h2>World Importer</h2>
 
 <form asp-controller="WorldImport" asp-action="Import" method="post">
-    <label asp-for="WorldDefinition"></label>  <textarea asp-for="Description"></textarea>
+    <label asp-for="WorldDefinition"></label>  <textarea asp-for="WorldDefinition.Description"></textarea>
     <button type="submit">Submit</button>
 </form>
